refactor: move squid to new dir
