--- src/commonMain/kotlin/exchange.dydx.abacus/processor/router/Squid/SquidTokenProcessor.kt
@@ -1,21 +1,0 @@
-package exchange.dydx.abacus.processor.router.Squid
-
-import exchange.dydx.abacus.processor.base.BaseProcessor
-import exchange.dydx.abacus.protocols.ParserProtocol
-
-internal class SquidTokenProcessor(parser: ParserProtocol) : BaseProcessor(parser) {
-    private val keyMap = mapOf(
-        "string" to mapOf(
-            "name" to "stringKey",
-            "address" to "type",
-            "logoURI" to "iconUrl",
-        ),
-    )
-
-    override fun received(
-        existing: Map<String, Any>?,
-        payload: Map<String, Any>
-    ): Map<String, Any> {
-        return transform(existing, payload, keyMap)
-    }
-}
