+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/router/squid/SquidProcessor.kt
@@ -1,17 +1,23 @@
-package exchange.dydx.abacus.processor.router.Squid
+package exchange.dydx.abacus.processor.router.squid
 
+import exchange.dydx.abacus.output.input.SelectionOption
+import exchange.dydx.abacus.output.input.TransferInputChainResource
+import exchange.dydx.abacus.output.input.TransferInputTokenResource
 import exchange.dydx.abacus.processor.base.BaseProcessor
-import exchange.dydx.abacus.processor.router.*
+import exchange.dydx.abacus.processor.router.IRouterProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
+import exchange.dydx.abacus.state.internalstate.InternalTransferInputState
 import exchange.dydx.abacus.state.manager.CctpConfig.cctpChainIds
-import exchange.dydx.abacus.utils.Logger
 import exchange.dydx.abacus.utils.mutable
 import exchange.dydx.abacus.utils.safeSet
 
-internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), IRouterProcessor {
-    override var chains: List<Any>? = null
-    override var tokens: List<Any>? = null
-    override var exchangeDestinationChainId: String? = null
+internal class SquidProcessor(
+    parser: ParserProtocol,
+    private val internalState: InternalTransferInputState,
+) : BaseProcessor(parser) {
+    private var chains: List<Any>? = null
+    private var tokens: List<Any>? = null
+    var exchangeDestinationChainId: String? = null
 
     override fun receivedChains(
         existing: Map<String, Any>?,
@@ -28,12 +34,12 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
         }
         val chainOptions = chainOptions()
 
-        modified.safeSet("transfer.depositOptions.chains", chainOptions)
-        modified.safeSet("transfer.withdrawalOptions.chains", chainOptions)
+        internalState.chains = chainOptions
+
         val selectedChainId = defaultChainId()
         modified.safeSet("transfer.chain", selectedChainId)
         selectedChainId?.let {
-            modified.safeSet("transfer.resources.chainResources", chainResources(selectedChainId))
+            internalState.chainResources = chainResources(selectedChainId)
         }
 
         updateTokensDefaults(modified, selectedChainId)
@@ -77,12 +83,12 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
         }
         val chainOptions = chainOptions()
 
-        modified.safeSet("transfer.depositOptions.chains", chainOptions)
-        modified.safeSet("transfer.withdrawalOptions.chains", chainOptions)
+        internalState.chains = chainOptions
+
         val selectedChainId = defaultChainId()
         modified.safeSet("transfer.chain", selectedChainId)
         selectedChainId?.let {
-            modified.safeSet("transfer.resources.chainResources", chainResources(selectedChainId))
+            internalState.chainResources = chainResources(selectedChainId)
         }
 
         updateTokensDefaults(modified, selectedChainId)
@@ -165,16 +171,15 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
     }
 
     override fun updateTokensDefaults(modified: MutableMap<String, Any>, selectedChainId: String?) {
-        val tokenOptions = this.tokenOptions(selectedChainId)
-        modified.safeSet("transfer.depositOptions.assets", tokenOptions)
-        modified.safeSet("transfer.withdrawalOptions.assets", tokenOptions)
+        val tokenOptions = tokenOptions(selectedChainId)
+        internalState.tokens = tokenOptions
         modified.safeSet("transfer.token", defaultTokenAddress(selectedChainId))
-        modified.safeSet("transfer.resources.tokenResources", tokenResources(selectedChainId))
+        internalState.tokenResources = tokenResources(selectedChainId)
     }
 
     override fun defaultChainId(): String? {
         val selectedChain = parser.asNativeMap(this.chains?.firstOrNull())
-        return parser.asString(selectedChain?.get("chain_id"))
+        return parser.asString(selectedChain?.get("chainId"))
     }
 
     override fun selectedTokenSymbol(tokenAddress: String?): String? {
@@ -186,7 +191,7 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
         return null
     }
 
-    override fun selectedTokenDecimals(tokenAddress: String?, selectedChainId: String?): String? {
+    override fun selectedTokenDecimals(tokenAddress: String?): String? {
         this.tokens?.find {
             parser.asString(parser.asNativeMap(it)?.get("address")) == tokenAddress
         }?.let {
@@ -228,44 +233,43 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
         }
     }
 
-    override fun chainResources(chainId: String?): Map<String, Any>? {
-        val chainResources = mutableMapOf<String, Any>()
+    override fun chainResources(chainId: String?): Map<String, TransferInputChainResource>? {
+        val chainResources = mutableMapOf<String, TransferInputChainResource>()
         chainId?.let {
             this.chains?.find {
                 parser.asString(parser.asNativeMap(it)?.get("chainId")) == chainId
             }?.let {
                 val processor = SquidChainResourceProcessor(parser)
                 parser.asNativeMap(it)?.let { payload ->
-                    chainResources[chainId] = processor.received(null, payload)
+                    chainResources[chainId] = processor.received(payload)
                 }
             }
         }
         return chainResources
     }
 
-    override fun tokenResources(chainId: String?): Map<String, Any>? {
-        val tokenResources = mutableMapOf<String, Any>()
+    override fun tokenResources(chainId: String?): Map<String, TransferInputTokenResource>? {
+        val tokenResources = mutableMapOf<String, TransferInputTokenResource>()
         filteredTokens(chainId)?.forEach {
             parser.asString(parser.asNativeMap(it)?.get("address"))?.let { key ->
                 val processor = SquidTokenResourceProcessor(parser)
                 parser.asNativeMap(it)?.let { payload ->
-                    tokenResources[key] = processor.received(null, payload)
+                    tokenResources[key] = processor.received(payload)
                 }
             }
         }
         return tokenResources
     }
 
-
-    override fun chainOptions(): List<Any> {
+    override fun chainOptions(): List<SelectionOption> {
         val chainProcessor = SquidChainProcessor(parser)
-        val options = mutableListOf<Any>()
+        val options = mutableListOf<SelectionOption>()
 
-        this.chains?.let { it ->
+        this.chains?.let {
             for (chain in it) {
                 parser.asNativeMap(chain)?.let { chain ->
-                    if (parser.asString(chain.get("chainType")) != "cosmos") {
-                        options.add(chainProcessor.received(null, chain))
+                    if (parser.asString(chain["chainType"]) != "cosmos") {
+                        options.add(chainProcessor.received(chain))
                     }
                 }
             }
@@ -275,26 +279,25 @@ internal class SquidProcessor(parser: ParserProtocol) : BaseProcessor(parser), I
         return options
     }
 
-
-    override fun tokenOptions(chainId: String?): List<Any> {
+    override fun tokenOptions(chainId: String?): List<SelectionOption> {
         val processor = SquidTokenProcessor(parser)
-        val options = mutableListOf<Any>()
+        val options = mutableListOf<SelectionOption>()
 
         val selectedChainId = chainId ?: defaultChainId()
         selectedChainId?.let {
+            selectedChainId
             this.tokens?.let {
                 for (token in it) {
                     parser.asNativeMap(token)?.let { token ->
-                        if (parser.asString(token.get("chainId")) == selectedChainId) {
-                            options.add(processor.received(null, token))
+                        if (parser.asString(token["chainId"]) == selectedChainId) {
+                            options.add(processor.received(token))
                         }
                     }
                 }
             }
         }
 
-        options.sortBy { parser.asString(parser.asNativeMap(it)?.get("stringKey")) }
+        options.sortBy { it.stringKey }
         return options
     }
-
 }
